Create app.rb and refactor main method in main.rb
