--- main
@@ -1,2 +1,0 @@
-#!/usr/bin/env ruby
-puts 'Hello!'
